Move HeterogeneousEvent to its own header
